@wip Upgrade to Serialport 4x
To be used in testing. By the 4.0.0 release I hope to have browser serialport in parity.

SerialPort 4.0.0-rc1 is out and brings performance improvements, bug fixes and a few changes to the api. It's now out of sync with `browser-serialport` but not in a major way. https://github.com/garrows/browser-serialport/issues/40

I suspect this upgrade won't have any far reaching effects as firmata uses only a fraction of the functions of SerialPort by default. The coms changes will be cleaner when browser serial port catches up.