chore: Upgrade SerialPort
SerialPort v8 brings support for the latest versions of node and electron. The api has had two major breaking changes. `.list()` no longer takes a callback and only returns a promise. And the `PortInfo` object now has `path` instead of `comName`.

This would probably be a breaking change for Firmata as well.